macOS: addressed a problem with the wallpaper not being updated.
The problem was due to macOS trying to be too smart by caching the file
name of our wallpaper. This means that even though we were modifying our
wallpaper, macOS (wrongly) assumed that nothing had changed since the
file name was still the same. Consequently, our new wallpaper wasn't
being set. A very simple workaround is therefore to use a different name
every time we need to update our wallpaper and we do this by adding a
timestamp to our file name (!!).