Now check every second that we are indeed showing the correct wallpaper.
This is related to the case where we switch desktops on, say, macOS.
Indeed, there doesn’t seem to be a way to know that we have switched
desktops, so we need to ensure that we are showing the correct
wallpaper on the ‘new’ desktop.

As a result, we reverted commit 937ce88 and thus only update the
wallpaper when actually necessary.